AIStockAgentMin/requirements.txt
fastapi==0.110.1
langchain_core==0.1.40
langchain_openai==0.1.1
langserve==0.0.51
python-dotenv==1.0.1
uvicorn==0.29.0
